Move tshirt_fluxes to dedicated header file.
Creating tshirt_fluxes.h and moving tshirt::tshirt_fluxes struct from
Tshirt.h to this new file.
