+++ models/tshirt/include/Tshirt.h
@@ -7,6 +7,7 @@
 #include "Nonlinear_Reservoir.hpp"
 #include "Pdm03.h"
 #include "GIUH.hpp"
+#include "tshirt_fluxes.h"
 #include "tshirt_params.h"
 #include "tshirt_state.h"
 #include <cmath>
@@ -15,29 +16,6 @@
 #include <memory>
 
 namespace tshirt {
-
-    /*!
-     * Tshirt flux structure
-     *
-     * This structure provides storage for the fluxes generated by Tshirt at any time step
-     */
-    struct tshirt_fluxes {
-        double surface_runoff_meters_per_second;  //!< Direct surface runoff, in meters per second
-        double groundwater_flow_meters_per_second;         //!< Deep groundwater flow from groundwater reservoir to channel flow
-        double soil_percolation_flow_meters_per_second;    //!< Percolation flow from subsurface to groundwater reservoir ("Qperc")
-        double soil_lateral_flow_meters_per_second;        //!< Lateral subsurface flow ("Qlf")
-        double et_loss_meters;         //!< Loss from ET, in meters
-
-        tshirt_fluxes(double q_gw = 0.0, double q_perc = 0.0, double q_lf = 0.0, double runoff = 0.0,
-                double et_loss = 0.0)
-                : groundwater_flow_meters_per_second(q_gw),
-                  soil_percolation_flow_meters_per_second(q_perc),
-                  soil_lateral_flow_meters_per_second(q_lf),
-                  surface_runoff_meters_per_second(runoff),
-                  et_loss_meters(et_loss) {
-
-        }
-    };
 
     // TODO: consider combining with or differentiating from similar hymod enum
     enum TshirtErrorCodes {
